fix: string build
