+++ src/lib/string.cpp
@@ -1,7 +1,5 @@
 #include "string.hpp"
 
-#include "nanoprintf.hpp"
-
 void memzero(void* start, void* end) {
   for (char* i = (char*)start; i != (char*)end; i++)
     *i = 0;
@@ -95,24 +93,4 @@ long strtol(const char* s, const char** endptr, size_t base, size_t n) {
   if (endptr)
     *endptr = s;
   return r * x;
-}
-
-bool operator==(string_view a, string_view b) {
-  if (a.size() != b.size())
-    return false;
-  return !memcmp(a.data(), b.data(), a.size());
-}
-
-string operator+(const string& a, const string& b) {
-  string s(a.size() + b.size() + 1);
-  s += a;
-  s += b;
-  return s;
-}
-
-string to_hex_string(uint64_t value) {
-  uint32_t size = npf_snprintf(nullptr, 0, "0x%lx", value) + 1;
-  string s(size);
-  s += npf_snprintf(s.data(), s.cap(), "0x%lx", value);
-  return s;
 }
